Change ArcWake::into_waker to a free function
